+++ src/layers/l2_norm.rs
@@ -8,7 +8,7 @@ use crate::{
 
 #[derive(Default, Debug, Clone)]
 pub struct L2NormCache {
-    pub x: Array2<f64>,
+    pub x: Array2<f32>,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
@@ -24,7 +24,7 @@ impl L2Norm {
         }
     }
 
-    pub fn forward(&mut self, x: Array2<f64>, grad: bool) -> Array2<f64> {
+    pub fn forward(&mut self, x: Array2<f32>, grad: bool) -> Array2<f32> {
         let l2_norm = f::l2_norm(&x);
 
         if grad {
@@ -34,7 +34,7 @@ impl L2Norm {
         l2_norm
     }
 
-    pub fn backward(&mut self, d_loss: Array2<f64>) -> Array2<f64> {
+    pub fn backward(&mut self, d_loss: Array2<f32>) -> Array2<f32> {
         f::d_l2_norm(&self.cache.x, &d_loss)
     }
 }
